fs: exec: Introduce oneplus_brain_service control
Based on concepts introduced in YaroST12/RAPHAEL_Kernel@a97f283

This provides an in-kernel mechanism to control the execution of
vendor.oneplus.hardware.brain@1.0-service. Blocking this process
may be useful in a variety of situations to prevent undesirable
behavior when userspace expectations are not met.

This alternative approach allows us to exercise control over the
process without relying on ramdisk manipulation or other root-based
solutions.

Co-authored-by: Yaroslav Furman <yaro330@gmail.com>
Co-authored-by: Adam W. Willis <return.of.octobot@gmail.com>